requirements.txt
streamlit
langchain
langchain_community
langchain_groq
langchain_core
mysql-connector-python
SQLALchemy
pymysql
chromadb
langchain-google-genai
python-dotenv==1.0.0
faiss-cpu
google-generativeai
openai==1.12.0
sqlite3==3.35.0
PyYAML==6.0.1